[FE] refactor : 메인 화면 리펙토링
