+++ gnuboardProject/frontend/src/App.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import Login from './components/Auth/Login';
 import Register from './components/Auth/Register';
-import PortfolioList from './components/Portfolio/PortfolioList';
 import PortfolioDetail from './components/Portfolio/PortfolioDetail';
 import PortfolioForm from './components/Portfolio/PortfolioForm';
 import Profile from './components/Profile/Profile';
 import AdminDashboard from './components/Admin/AdminDashboard';
 import { useAuth } from './contexts/AuthContext';
 import PortfolioWrite from './components/Portfolio/PortfolioWrite';
+import Main from './components/Portfolio/Main';
 
 const PrivateRoute: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const { isAuthenticated } = useAuth();
@@ -26,7 +26,7 @@ const App: React.FC = () => {
       <Routes>
         <Route path="/login" element={<Login />} />
         <Route path="/register" element={<Register />} />
-        <Route path="/" element={<PortfolioList />} />
+        <Route path="/" element={<Main />} />
         <Route path="/portfolios/:id" element={<PortfolioDetail />} />
         <Route
           path="/portfolios/new"
